Implemented recent optimizations in GPU and moved features to relevant modules
