+++ src/spatial/raytracing/mod.rs
@@ -1,4 +1,4 @@
-use crate::spatial::{math::vector::V3c, Cube};
+use crate::spatial::{math::vector::V3c, raytracing::lut::OCTANT_STEP_RESULT_LUT, Cube};
 
 pub mod lut;
 mod tests;
@@ -59,4 +59,71 @@ impl Cube {
             impact_distance: Some(tmin),
         })
     }
+}
+
+/// Provides the resulting octant based on the given octant and a direction it is stepping to
+/// It returns with 8 if the resulting octant is out of bounds.
+/// Important note: the specs of `signum` behvaes differently for f32 and i32
+/// So the conversion to i32 is absolutely required
+pub(crate) fn step_octant(octant: u8, step: V3c<f32>) -> u8 {
+    let octant_pos_in_32bits = 4 * octant;
+    ((OCTANT_STEP_RESULT_LUT[((step.x as i32).signum() + 1) as usize]
+        [((step.y as i32).signum() + 1) as usize][((step.z as i32).signum() + 1) as usize]
+        & (0x0F << octant_pos_in_32bits))
+        >> octant_pos_in_32bits) as u8
+}
+
+/// calculates the distance between the line, and the plane both described by a ray
+/// plane: normal, and a point on plane, line: origin and direction
+/// return the distance from the line origin to the direction of it, if they have an intersection
+#[allow(dead_code)] // Could be useful either for debugging or new implementations
+pub fn plane_line_intersection(
+    plane_point: &V3c<f32>,
+    plane_normal: &V3c<f32>,
+    line_origin: &V3c<f32>,
+    line_direction: &V3c<f32>,
+) -> Option<f32> {
+    let origins_diff = *plane_point - *line_origin;
+    let plane_line_dot_to_plane = origins_diff.dot(plane_normal);
+    let directions_dot = line_direction.dot(plane_normal);
+    if 0. == directions_dot {
+        // line and plane is paralell
+        if 0. == origins_diff.dot(plane_normal) {
+            // The distance is zero because the origin is already on the plane
+            return Some(0.);
+        }
+        return None;
+    }
+    Some(plane_line_dot_to_plane / directions_dot)
+}
+
+pub fn cube_impact_normal(cube: &Cube, impact_point: &V3c<f32>) -> V3c<f32> {
+    let mid_to_impact =
+        V3c::from(cube.min_position) + V3c::unit(cube.size as f32 / 2.) - *impact_point;
+    let max_component = mid_to_impact
+        .x
+        .abs()
+        .max(mid_to_impact.y.abs())
+        .max(mid_to_impact.z.abs());
+
+    let impact_normal = V3c::new(
+        if mid_to_impact.x.abs() == max_component {
+            -mid_to_impact.x
+        } else {
+            0.
+        },
+        if mid_to_impact.y.abs() == max_component {
+            -mid_to_impact.y
+        } else {
+            0.
+        },
+        if mid_to_impact.z.abs() == max_component {
+            -mid_to_impact.z
+        } else {
+            0.
+        },
+    );
+
+    debug_assert!(0. < impact_normal.length());
+    impact_normal.normalized()
 }
